.gitignore updates, add examples/Makefile "local"
Miscellaneous updates to .gitignore files and a new "local" target to
the examples Makefile. This Makefile will now build versions of the
example programs that explicitly reference the local repository version
of FANN (as opposed to a global one in /usr/lib or otherwise). This is
useful if you have multiple versions of FANN or want to make some local
modifications without changing whatever is globally available.